sim(64-bit): make engine/storage/model pointer math 64-bit-clean
Apply the right expression per intent across the audio engine, wavetable, storage and
model code: pointer differences -> native subtraction; byte-offset-into-buffer -> char*
arithmetic; address low-bit inspection -> uintptr_t; tagged Drum*/Clip* fields carrying an
encoded 32-bit index during load -> read via (uintptr_t) (avoids a truncated real pointer
matching a sentinel); the input-ring read cursor inputRingPos becomes uintptr_t (it held an
address in a uint32_t); and std::min(numSamplesFeedingNow, 256u) -> size_t{256} to match the
size_t operand on the 64-bit host. Behaviour is unchanged on the 32-bit firmware.

+++ src/deluge/processing/engines/audio_engine.h
@@ -198,7 +198,7 @@ extern bool bypassCulling;
 // sample recorders) reads this, never the hardware DMA ring. Same geometry as
 // the board's RX ring: SSI_RX_BUFFER_NUM_SAMPLES frames of
 // NUM_MONO_INPUT_CHANNELS interleaved q31 samples.
-extern uint32_t inputRingPos; // read cursor: input aligned with the next output sample
+extern uintptr_t inputRingPos; // read cursor (an address): input aligned with the next output sample
 int32_t* inputRingStart();
 int32_t* inputRingEnd();
 extern int32_t cpuDireness;
